// src/model/nation_wide_rating.ts
export interface ModelNationWideRating {
    candidate1Id: number
    candidate2Id: number
    candidate1Name: string
    candidate2Name: string
    candidate1Img: string
    candidate2Img: string
    value: number
    trust: number
    anger: number
    anticipation: number
    disgust: number
    fear: number
    joy: number
    sadness: number
    surprise: number
}